FOEPRO/requirements.txt
Django==5.2.10
psycopg2-binary
